Implement getNextUnqiueId using callbacks
This solution was done carefully with attention to solution video with annotations.

*pair was largely unresponsive to suggestions for direction and problem solving process (including identifiying the problem) for hours, then was suddenly far ahead the next morning.

Working backwards left me lost on the idea of callbacks, so I decided to start fresh and approach the problem carefully on my own and within time constraints.

These are new commits according to that on my master branch.